feat: enhance query handling with cluster support and improve UI components
