+++ app/[host]/query/[query_id]/page.tsx
@@ -12,6 +12,7 @@ export const revalidate = 300
 
 export default async function Page({ params, searchParams }: PageProps) {
   const { query_id } = await params
+  const { cluster } = await searchParams
 
   // Binding the query_id to the config
   const queryConfig = {
@@ -21,6 +22,13 @@ export default async function Page({ params, searchParams }: PageProps) {
     },
   }
 
+  if (cluster) {
+    queryConfig.sql = queryConfig.sql.replace(
+      'FROM system.query_log',
+      `FROM clusterAllReplicas('${cluster}', system.query_log)`
+    )
+  }
+
   return (
     <div className="flex flex-col gap-4">
       <Suspense fallback={<ChartSkeleton />}>
@@ -28,7 +36,11 @@ export default async function Page({ params, searchParams }: PageProps) {
       </Suspense>
 
       <Suspense fallback={<TableSkeleton />}>
-        <QueryDetail queryConfig={queryConfig} params={await params} />
+        <QueryDetail
+          queryConfig={queryConfig}
+          params={await params}
+          searchParams={await searchParams}
+        />
       </Suspense>
 
       <Suspense fallback={<TableSkeleton />}>
