+++ app/[host]/query/[query_id]/types.ts
@@ -2,5 +2,5 @@ export interface PageProps {
   params: Promise<{
     query_id: string
   }>
-  searchParams: Promise<{ [key: string]: string | string[] | undefined }>
+  searchParams: Promise<{ cluster?: string }>
 }
